templates fix and generator update

+++ src/generator.ts
@@ -1,35 +1,60 @@
 import fs from "fs-extra";
 import path from "path";
 import ejs from "ejs";
+import { Layer } from "./layers/Layer.enum";
+import { toPascalCase } from "./utils/toPascalCase";
+import chalk from "chalk";
 
 const TEMPLATES_DIR = path.join(__dirname, "..", "templates", "slice");
 
-export const generateSlice = async (sliceName: string) => {
-  const sliceDir = path.join(process.cwd(), "src", "slices", sliceName);
+export const generateSlice = async (
+  layer: Layer,
+  name: string,
+  opt?: string
+) => {
+  const sliceDir = path.join(process.cwd(), "src", layer, name);
+  if (await fs.pathExists(sliceDir)) {
+    console.error(
+      chalk.red(
+        `Error: slice with name '${name}' already exists on layer '${layer}'.`
+      )
+    );
+    process.exit(1);
+  }
 
   try {
     await fs.ensureDir(sliceDir);
-    await fs.ensureDir(path.join(sliceDir, "components"));
 
     const indexTemplate = await fs.readFile(
       path.join(TEMPLATES_DIR, "index.ts.ejs"),
       "utf-8"
     );
-    const indexContent = ejs.render(indexTemplate, { sliceName });
+    const indexContent = ejs.render(indexTemplate, {
+      name: toPascalCase(name),
+      opt,
+    });
     await fs.writeFile(path.join(sliceDir, "index.ts"), indexContent);
 
     const componentTemplate = await fs.readFile(
       path.join(TEMPLATES_DIR, "component.tsx.ejs"),
       "utf-8"
     );
-    const componentContent = ejs.render(componentTemplate, { sliceName });
+    const componentDir = path.join(sliceDir, "ui");
+    await fs.ensureDir(componentDir);
+    const componentContent = ejs.render(componentTemplate, {
+      name: toPascalCase(name),
+    });
     await fs.writeFile(
-      path.join(sliceDir, "components", `${sliceName}Component.tsx`),
+      path.join(componentDir, `${toPascalCase(name)}.tsx`),
       componentContent
     );
 
-    console.log(`Слайс '${sliceName}' успешно создан!`);
+    console.log(
+      chalk.green(
+        `OK: Slice '${name}' layer '${layer}' has been successfully created`
+      )
+    );
   } catch (error) {
-    console.error("Ошибка при создании слайса:", error);
+    console.error(chalk.red("Error: Slice creating error:", error));
   }
 };
